Simplify: extract detectAndSaveUserLocation, consolidate Coords
Both detect-location callsites (settings page + ReferenceLocationInput)
were running the same 5-step pipeline (getCurrentPosition → apiGeocode
→ label fallback → saveUserLocation → distinguish denied). Extract
into detectAndSaveUserLocation() so the two callers only handle their
UI state. Consolidate the Coords interface — was duplicated between
lib/geolocation.ts and lib/userProfile.ts.UserLocation; now Coords
lives in userProfile and UserLocation extends it.

Centralizes default options (timeout: 10000, enableHighAccuracy: false)
in one constant, drops the dead intermediate destructuring after the
plugin null-guard.

Documents the iOS WKWebView geolocation pitfall + the canonical
lib/geolocation.ts entry point in CLAUDE.md.
