+++ CLAUDE.md
@@ -1854,6 +1854,18 @@ Per-group "New Poll" notifications via Web Push (browser / PWA / iOS PWA 16.4+) 
 - **`POST /api/notifications/subscriptions` is upsert-on-(browser_id, endpoint).** Re-subscribing the same browser refreshes encryption keys (Web Push rotates them) AND resets `failure_count`/`last_error`. Don't try to dedupe client-side based on stored subscription state — let the server's ON CONFLICT handle it.
 - **Adding a new npm dep requires `npm install --package-lock-only`** to regenerate `package-lock.json` before pushing — the Mac dev-server entrypoint uses `npm ci` which fails hard on lockfile drift. Symptom: dev container loop-restarts with `npm error` referencing `npm ci` not matching `package.json`. Same lockfile must be in the commit. Server-side `uv.lock` is regenerated via `uv lock` from `server/pyproject.toml`; the Dockerfile uses `uv sync --frozen` IF lock exists so it must be committed alongside the dep change.
 
+## Geolocation (Native + Web)
+
+`lib/geolocation.ts` is the canonical entry point for "where is the user?" — every callsite must route through it, not directly through `navigator.geolocation`.
+
+- **WKWebView in Capacitor silently denies `navigator.geolocation`.** Capacitor's `CAPBridgeViewController` doesn't implement `WKUIDelegate.webView(_:requestGeolocationPermissionFor:initiatedBy:decisionHandler:)`, so a JS call to `navigator.geolocation.getCurrentPosition` in the iOS app produces no permission prompt, no error, and no position — the callback just never fires (or fires the error callback with a generic POSITION_UNAVAILABLE depending on iOS version). Adding `NSLocationWhenInUseUsageDescription` to Info.plist is necessary but NOT sufficient — the plist key enables the OS prompt, but without a delegate to receive the WebView's geolocation request, the prompt never shows. The fix is the `@capacitor/geolocation` plugin, which bridges to native Core Location.
+- **`getCurrentPosition(options?)` dispatches to native or web.** Under `Capacitor.isNativePlatform()` it dynamic-imports `@capacitor/geolocation` and routes through `Geolocation.checkPermissions() → requestPermissions() → getCurrentPosition()`. On web (browsers / PWA / iOS PWA) it falls back to `navigator.geolocation`. Both paths return the same `Coords` shape (`{latitude, longitude}` from `lib/userProfile.ts`) and throw a unified `GeolocationDeniedError` when permission is refused so callers can `instanceof`-branch the catch instead of string-matching.
+- **`detectAndSaveUserLocation()` is the full pipeline** — calls `getCurrentPosition()`, reverse-geocodes via `apiGeocode(...)`, builds a label fallback, and persists via `saveUserLocation(...)`. Both detect-location callsites (`components/ReferenceLocationInput.tsx`, `app/settings/page.tsx`) consume this helper; differences are confined to UI state (input clear, `setMessage` vs `setError`). When adding a third "detect my location" surface, reuse this helper rather than re-doing the pipeline.
+- **`NSLocationWhenInUseUsageDescription` MUST be in `ios/App/App/Info.plist`** with a justification string the user will actually see in the permission prompt. iOS rejects the app at first geolocation request with no prompt at all if the key is missing. No matching entitlement is needed — `aps-environment` and `com.apple.developer.associated-domains` are separate capabilities; geolocation has no entitlement.
+- **`Coords` lives in `lib/userProfile.ts`, and `UserLocation extends Coords`.** Don't redefine the lat/lng pair locally — import from `userProfile.ts`. The label (city/zip display string) is the only thing `UserLocation` adds on top.
+- **Native dispatch uses the dynamic-import-Capacitor-plugin pattern** (`await import("@capacitor/...").catch(() => null)`), matching `lib/pushNotifications.ts: ensureCapacitorPushSubscription` and `lib/universalLinks.ts: installUniversalLinksHandler`. The chunk only loads on the native bundle; web bundles include the chunk but never execute it (the `Capacitor.isNativePlatform()` short-circuit returns false). Don't `import` the plugin statically at module top — that pulls it into every web bundle for no benefit.
+- **Three plugins now use this pattern.** If a fifth lands, extract a shared `loadCapacitorPlugin<T>(name, exportKey)` helper; the current ~4-line repetition is below the abstraction threshold.
+
 ## App Icons
 
 The 👋 waving-hand emoji is the canonical app icon, rendered on a black rounded-rect background. Every icon file in the project is a render of the same SVG template — keep them in sync when changing the artwork.
